--- benchmark/FileOp.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <fstream>
-#include <vector>
-#include <iostream>
