--- benchmark/Parser.h
@@ -1,1 +1,0 @@
-
