--- benchmark/ScopeExit.h
@@ -1,1 +1,0 @@
-
