+++ benchmark/cupti_test.h
@@ -10,9 +10,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-
-#include "Eval.h"
 #include "Utils.h"
+//#include "Eval.h"
+
 
 #ifndef EXIT_WAIVED
 #define EXIT_WAIVED 2
